backend/requirements.txt
Django==5.2.6
python-dotenv
pytest
pytest-cov
djangorestframework==3.16.1
djangorestframework-simplejwt>=5.3.0